script modified to not erase gsmt2 file when using 'only-translate' flag

+++ scripts/AllRepair.sh
@@ -26,7 +26,7 @@ opts="ufmntsph:"
 # always clean before exit
 clean(){
 	# delete files created by script
-	if [[ -f $creating_file ]] && [[ $KEEP -ne 1 ]]; then
+	if [[ -f $creating_file ]] && [[ $KEEP -ne 1 ]] && [[ $REPAIR -ne 0 ]]; then
 		#echo "Last file created: $creating_file";
 		rm $creating_file
 	fi
